fix the size_t error

+++ include/common/string.h
@@ -7,6 +7,6 @@ void *memset(void *str, int c, size_t n);
 void *memcpy(void *dst_str, const void *src_str, size_t n);
 size_t strlen(const char *str);
 int strcmp(const char *str1, const char *str2);
-int strcmp_s(const char *str1, const char *str2, int n);
+int strcmp_s(const char *str1, const char *str2, size_t n);
 
 #endif
